Extract NavItems into separate file
